requirements.txt
Django==3.2.7

django-cors-headers==3.10.0

mysqlclient==2.0.3

pyjwt==2.3.0

requests==2.26.0

bcrypt ==3.2.0